# Conflicts:
#	LurkingMonster/Assets/0. Scenes/Game.unity
#	LurkingMonster/Assets/1. Scripts/Gameplay/HouseBreak.cs

+++ LurkingMonster/Assets/1. Scripts/Events/CollectRentEvent.cs
@@ -4,9 +4,9 @@ namespace Events
 {
 	public class CollectRentEvent : VDEvent
 	{
-		public readonly float Rent;
+		public readonly int Rent;
 
-		public CollectRentEvent(float rent)
+		public CollectRentEvent(int rent)
 		{
 			Rent = rent;
 		}
